+++ app/src/main/java/fr/free/nrw/commons/di/ActivityBuilderModule.java
@@ -13,7 +13,7 @@ import fr.free.nrw.commons.contributions.MainActivity;
 import fr.free.nrw.commons.category.CategoryImagesActivity;
 import fr.free.nrw.commons.explore.SearchActivity;
 
-import fr.free.nrw.commons.nearby.WikiFeedback;
+import fr.free.nrw.commons.nearby.WikidataFeedback;
 import fr.free.nrw.commons.explore.categories.ExploreActivity;
 import fr.free.nrw.commons.notification.NotificationActivity;
 import fr.free.nrw.commons.settings.SettingsActivity;
@@ -66,6 +66,6 @@ public abstract class ActivityBuilderModule {
     abstract BookmarksActivity bindBookmarksActivity();
 
     @ContributesAndroidInjector
-    abstract WikiFeedback bindWikiFeedback();
+    abstract WikidataFeedback bindWikiFeedback();
 
 }
