+++ app/src/main/res/layout/activity_wiki_feedback.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".nearby.WikiFeedback">
+    tools:context=".nearby.WikidataFeedback">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/textHeader"
